Simplify completions
Turns out I was doing a lot of the work that bash already does
for me. I don't need to filter the lists or sort them. Bash will
do that for me. So the complete for set and member methods are
much simpler now. They just return all set names (for sets) or
all member names (for sets). In addition to names all of the relevant
ids are returned as well.