Switch to using just one data file
The new data file has all sets, and contains original
names and the new camel case names. And the layout is
much more client friendly.

Added memoization to getEnums and getEnumSet so that
if this is later used as a library it will have faster
access after a set has been retrieved previously.

This makes for a smaller package size (less than half size
due to less duplication)