add logic in santity check to generate no dupes file
When test runs it finds duplicate ids for some members
because the original logic for generating allSets.enum didn't
take care of this. So now the sanity-check test script will generate
a file that repairs duplicates the same way the production process does.
Then this can be checked in instead of the version with dupes.